## Description

This PR's diff is much smaller than it looks! There's an embedded XML resource, and adjusting this meant that the indentation levels changed. When that's accounted for, the PR becomes much smaller - around 300 lines, rather than 7800.

This PR relates to #1942 and #1261. At present, there are two identical copies of the `Microsoft.Data.SqlClient.SqlMetaData.xml` embedded resource in the netfx and the netcore projects. This PR merges these with a single file. More importantly though, this embedded resource is used only within `DbMetaDataFactory`; it's deserialized using `DataSet.ReadXml`, where it is eventually used in `SqlConnection.GetSchema`.

`DataSet.ReadXml` isn't trim-safe - it uses XML serialization in the background. I've replaced this with a simple parsing routine which runs against the embedded resource and adds the rows manually.

The net result is smaller, faster and more efficient. Sizoscope confirms that this PR cuts the file size of a simple application by 3.8MB, from 20.4MB to 16.6MB. Benchmark results for this in isolation are:

| Method               | Mean       | Error    | StdDev   | Ratio | RatioSD | Gen0     | Gen1    | Allocated | Alloc Ratio |
|--------------------- |-----------:|---------:|---------:|------:|--------:|---------:|--------:|----------:|------------:|
| Before | 1,819.7 μs | 25.07 μs | 23.45 μs |  1.00 |    0.02 | 187.5000 | 39.0625 | 829.81 KB |        1.00 |
| After |   738.9 μs |  3.25 μs |  2.71 μs |  0.41 |    0.01 |  62.5000 | 19.5313 | 307.74 KB |        0.37 |

Separately, I noticed that my earlier PR #3309 generated two warnings when publishing because it wasn't using a fully-qualified type name. I've adjusted this.

## Issues

Relates to #1942.
Relates to #1261.
Follows up #3309.

## Testing

All tests continue to pass.

There's a lot of validation of the structure of the XML in the embedded resource, and this is done with debug assertions. I've treated this as hardcoded at the point of build, so haven't used exceptions for this. I can change this if needed.